Use index as value in StackTracesTreeNode maps.
Summary:
`StackTracesTreeNode` stores various `llvm::DenseMap` which allow it to point
to its children `StackTracesTreeNode *`s.

The need to iterate these children resulted in the creation of
a `ChildrenIterator` which iterated in an unspecified order.

Remove the nondeterminism incurred by iterating `llvm::DenseMap` by
storing all child nodes in an `std::vector` and using indexes as
values in the `llvm::DenseMap`. This also simplifies iteration as we
can iterate `.getChildren()` directly.

The other option was to use `std::map`, but it doesn't provide the same
`try_emplace` API as `llvm::DenseMap` and this is a simpler fix with a
cleaner iteration method.

Reviewed By: dulinriley

Differential Revision: D21505460

fbshipit-source-id: be1ca655eae43f070f9c1d00a875a7fe4ef6182d