Avoid memsetting new slabs in BacktrackingBumpPtrAllocator.
Summary:
`allocateNewSlab` was using `llvm::make_unique`, which calls the
constructor on the template parameter, and forces a `memset` on the
whole new slab.

We don't care if the new slab is zero-initialized because we're doing
bump pointer allocation and we will overwrite the allocated memory with
something useful any way.

Use `new Slab` without parentheses and avoid `make_unique` to prevent
`memset` from running.

Reviewed By: tmikov

Differential Revision: D21575586

fbshipit-source-id: 513587f75e092007d4d966477f2f23816b75bf84