Add Intl object to JSLib
Summary:
This includes the interface between the common JSLib Intl
code and the platform-specific code, but no implementations of that
code.  So, it would compile all the source files if enabled, but will
not link.

This does not provide support for all of ECMA 402.  It only supports APIs which are in JSC: Collator, DateTimeFormat, and NumberFormat, and not DateTimeFormat's range methods.  PluralRules is supported by JSC, but not here.  I chose to leave it for later, since it was implemented recently, and I wanted to get the key functionality in first.

Reviewed By: avp

Differential Revision: D21420122

fbshipit-source-id: 9a492fc657c6ea993ecde1e36a74c86e76c0288e