Update write barriers for Hades
Summary:
The write barrier API for other GCs (namely GenGC) was agnostic to some details,
such as the following:
* The order of the write barrier and the write to memory
* The written-to location's previous value (or lack thereof in `uninitialized` cases)

These details become more important in other GCs like Hades.
So, to be conservative, have all users of write barriers use the Hades-centric
API.

This involves a couple changes:
* Requiring that the write barrier occurs before the write to memory.
This is important to make guarantees when marking is made concurrent,
and these barriers are interleaved with GC actions
* Adding a different write barrier in the case where a `GCPointer` or
`GCHermesValue` is constructed, compared to when an existing pointer
is mutated

These changes are represented by a new `constructorWriteBarrier` API, and by
re-ordering the calls to the API.

As Hades stands right now, this doesn't make a material difference; however,
in a follow-up diff that adds concurrent marking, this difference is important
(and there will be different code for the `constructorWriteBarrier` than the
normal write barrier).

Reviewed By: davedets

Differential Revision: D20526991

fbshipit-source-id: 0996edff552e95bc94a8fe2a60e255456550f95f