Factor out logic to statically link ICU
Summary:
This diff adds the ability to specify that only libicu should be
statically linked. It also defines `U_STATIC_IMPLEMENTATION`
when libicu is being statically linked, which was previously
missing.

In addition, while testing this, I found that the usual static build
was a little broken because it is not possible to build libhermes.so
with the statically linked icu provided by package managers
since they are compiled without `-fPIC`. This diff therefore also
adds a check to `hermes_link_icu` that throws an error when
a shared object attempts to link against a static ICU.

Reviewed By: tmikov

Differential Revision: D21191935

fbshipit-source-id: 39df6f94479d4d292cebd7d8a5c2f8284cc91a0c