Ignore empty type fields in ESTreeJSONDumper.
Summary:
As we add support for parsing type annotations to the Hermes parser,
we must add new properties to many of the existing nodes.

To avoid adding lots of noise to diffs and extra useless lines to
test files, we introduce a mechanic to allow us to mark certain fields
on certain ESTree nodes as "ignore when empty".

These fields won't be dumped by `-dump-ast` when they are set to `nullptr`.

We add existing fields such as `returnType` and `typeAnnotation`
to this category, in order to maintain consistency as we move forward
and ignore empty type fields in subsequent diffs. This adds noise to this
diff, but only this diff.

Reviewed By: tmikov

Differential Revision: D21391073

fbshipit-source-id: 62d9dcbf9b4ba8746c93d357d7656798f6db591c