Set recursion depth limit on regex parser
Summary:
Hermes regex parsing currently uses recursion to process alternations
and capture groups. This diff imposes a maximum limit on the depth of
that recursion.

Note that this limit applies to both the maximum
stack depth in the parser and the maximum nesting depth of the data
structures it produces. It therefore also protects later stages
(optimisation and codegen) from stack overflows.

Reviewed By: avp

Differential Revision: D21822444

fbshipit-source-id: 5327b61e38a42865fbca1ca82546f58aead07f4e