Add weak ref read barrier
Summary:
In a concurrent GC like Hades, there is a problem with a
weak reference: reading from it creates a strong pointer to something
that is otherwise unreachable.

Depending on where that value is stored, and how marking is implemented,
there's no guarantee that the now-reachable value will be kept alive.

When a weak reference is read, be conservative and assume that it might get
written somewhere where it will stay alive. So mark it as live there. In the worst case,
this could result in a weak reference staying around indefinitely if it happens to
get read during every collection.

This diff adds a read barrier to all `WeakRef` instances.
This read barrier currently does nothing on GenGC, and will do something
in Hades in a subsequent diff that adds concurrent marking.
It also introduces the `WeakRoot<T>` type, which does read barriers for weak roots.

A second problem with concurrent marking of weak references is that they
live in off-heap data structures like `llvm::DenseMap` that can move or resize
while marking is ongoing. The mutator will share a `WeakRefMutex` with
the GC to make sure that accesses to these data structures are protected.
To enforce this, the read barrier APIs all check the mutex.
This mutex can go away if those external structures are moved
to become on-heap. The mutex is taken by all APIs to ensure no accidents happen
with WeakRef-containing code.

Through the writing of this diff, I found that `WeakRefHolder` and weak ref support
in `HandleRootOwner` were unnecessary, so I deleted them to save some complexity
of supporting a use case that didn't need to exist.

Reviewed By: davedets

Differential Revision: D20526993

fbshipit-source-id: 9f63e7e6872ea7b67a28af1b009d967eae83380c