Add a flag for optimized eval
Summary:
Sometimes it might be desired to run compiler optimizations on the JS source provided as an
argument to `eval`. For instance:
* If a program is mostly composed of large `eval` statements
* If we wanted to test the optimizer via `eval` statements

So add a runtime flag (that defaults to off) allowing the control of this behavior.

When the flag is set, lazy compilation is always disabled.

This change brought out a bug in DCE, which could accidentally delete the entry point of the module.
Add an explicit check for cases when this is not also global scope.

Reviewed By: dulinriley

Differential Revision: D15329403

fbshipit-source-id: 276e90d40428ccf546b8db595bb0f78056b43e70