Change runtime error message for invalid RegExp
Summary:
As part of consolidating validation of the regex flags, the error
returned should now be general enough to indicate that there was an
error in either the pattern or the flags. A follow up error will then
describe exactly what went wrong.

This updates error messages and tests in preparation for that
change.

Reviewed By: avp

Differential Revision: D21978017

fbshipit-source-id: d5092d07d619102536dc4ff135aac899121d13aa