Use SyntaxFlags struct when compiling regex
Summary:
This diff changes the underlying implementation of the regex parser
and compiler to use the new SyntaxFlags struct as opposed to the
constants::SyntaxFlags enum.

It also adds the fromByte and toByte functions to convert between
the struct and the bytecode representation. The offsets used match
the ones specified in the enum.

This is better because:
1. The code is now consistent across compilation and runtime so there
is no need to convert between the two.
2. The types are easier to manage without worrying about casting
between `uint8_t` and `SyntaxFlags`.
3. All the flag code is now in a single place.

A subsequent change can update the API to take a string and use the
`fromString` function to convert and validate the input internally in the
compiler.

Reviewed By: avp

Differential Revision: D21977059

fbshipit-source-id: 5c19e27e13ef3299fa899a75911dbc7759a9e90f