Factor out flag validation
Summary:
The FlagBits struct indicates which regex flags have been specified
and is currently only used at runtime.

This change moves the FlagBits struct and the regex flag validation
code into Compiler.h so that it is accessible to the regex compiler.
It also renames the struct to `SyntaxFlags` for consistency.

Subsequent changes can start using this struct in the compiler so
that the validation code can be shared and the SyntaxFlags enum
can be removed.

Reviewed By: avp

Differential Revision: D21977056

fbshipit-source-id: 43c3e467b9116416788f69493e128b5a2c4c380e