Have the SamplingProfiler wait for the timer thread to exit on disable
Summary:
If the main thread exits before the SamplingProfiler's timer thread shuts itself down,
the timer thread could read from freed memory (including itself, which lives in a static
variable).

Now, whenever the SamplingProfiler is disabled, it explicitly waits for its timer thread to complete.
The timer thread also checks `enabled_` whenever it releases the profiler lock in case it was changed,
so that it exits early.

Reviewed By: tmikov

Differential Revision: D22189475

fbshipit-source-id: 327ed9f68c0e3cb1b81866bac036ae5a5b91ab38