Add support for numeric literal separators.
Summary:
Numeric literal separators are a stage 3 proposal:
https://tc39.es/proposal-numeric-separator/

Numeric literals may now include `_` between any two digits.
The `_` is prohibited at the start or end of a numeric literal,
or adjacent to the `.`, `e`, etc.

We avoid incurring too much extra cost here by setting a `seenSeparator`
flag on the first pass through the number before we calculate its value.
In the vast majority of cases, there is no separator and we can use our
usual methods. In particular, the fast pass for <9-digit decimal integers
only incurs one extra check in the branch.

Anything involving the separator is then delegated to `strtod` or to
`parseIntWithRadix`. The former involves copying the characters that
are not `_` to a buffer to pass to the third-party function, while
the latter has been augmented with a flag to enable ignoring `_`
(other users such as `Number()` must still error on seeing `_`).

Reviewed By: tmikov

Differential Revision: D21796446

fbshipit-source-id: a35c0c493224033647235254811eed5688143771