Parse Flow enums.
Summary:
Flow has support for `enum` declarations, which are fairly straightforward
to parse, but have a few typechecking cases which need to be handled
in order to generate a valid AST, as the node names of the enum bodies
are themselves typed (e.g. `EnumStringBody`, `EnumNumberBody`, etc).

Add support for parsing the various flavors of enum, as well
as exporting said enums using `export` and `export default`.

Reviewed By: tmikov

Differential Revision: D21239327

fbshipit-source-id: 853ffda0a253fc4e1a6493aa223a9b5a687de2b2