Fix string split matches at the end of strings
Summary:
The spec requires that we can only match on separators when the match
starts at a position  before the end of a string. Our current
implementation allows matches that start after the last character.
This is incorrect because, for instance, we shouldn't be able to match
and split with `/$/`.

Reviewed By: avp

Differential Revision: D22225720

fbshipit-source-id: 4ad7c1fa6a88ccaef869ee664e2625bedd100436