Implement Object.getOwnPropertyDescriptors.
Summary:
Standard JSLib function that collects all the property descriptors
for a given object.

The current implementation is not fully optimized, as we don't have a
callback-style API for enumerating all properties in an object including
indexed properties, so we use an intermediate array to store them.
This can be improved upon in the future, but this diff provides
the functionality necessary to be standards-compliant.

Reviewed By: tmikov

Differential Revision: D22140163

fbshipit-source-id: 536ef66d76e0e9b7b8084d8304f4d2b597e386af