Fix WeakRef and Symbol allocate live bug in Hades
Summary:
This diff fixes two bugs in Hades:
* Newly created WeakRefs were not getting marked, and got cleared improperly
* Symbols were actually being initialized to **not** marked accidentally,
because the constructor was run **after** the call to `entry.mark()`

The fix for the latter involves having all Symbols marked by default, and I realized
that we can clear the marked bit at the beginning of the collection phase to free
more symbols. I also realized that utilizing a STW pause to free symbols means
there's no race condition when allocating symbols. So I deleted the lookupVectorMutex_
and the firstGCFreeID_ code that is now unnecessary.

These bugs were found while I was implementing Handle-SAN for Hades.

Reviewed By: davedets

Differential Revision: D22326798

fbshipit-source-id: fc8a42ae7f6b6903a86938ea973c9c5e46e5caee