Fix implicit int/float conversion warning in JSLib/ArrayBuffer
Summary:
There was a compiler warning from clang about an implicit lossy conversion
from int to double.

```
hermes/lib/VM/JSLib/ArrayBuffer.cpp:102:20:
warning: implicit conversion from 'unsigned long' to 'double' changes value
from 18446744073709551615 to 18446744073709551616.
[-Wimplicit-int-float-conversion]
```

It happens on the comparison: `byteLength > std::numeric_limits<JSArrayBuffer::size_type>::max()`,
since `byteLength` is a double.
Since byteLength is the result of a call to `ToIndex`, we know it is an unsigned integer anyway, so
store it as a `uint64_t` instead of a double to prevent the warning.

Reviewed By: avp

Differential Revision: D22399958

fbshipit-source-id: 3fad4852e1dda65edba5ca14d3436ffc0f04475b