Don't hold the WeakRef mutex while calling alloc
Summary:
There was a potential deadlock opportunity if a caller tried
to allocate memory while wolding the WeakRef mutex,
as a young gen collection would attempt to acquire it as well.

Add a slow assert to ensure that the mutex isn't held.
It's a slow assert because alloc is a hot path, and checking the
mutex's owner requires accessing the current thread id.

Also fix one location where this issue occurred, in `WeakMap`.

Reviewed By: davedets

Differential Revision: D22417637

fbshipit-source-id: 53265ada5a26a6249f24b711175bbe0679dc99a9