Initialize the new cell in allocLongLived to prevent the sweeper from reading it
Summary:
There was a rare edge case in allocating directly into the old gen:
if it happened while a segment was getting swept, and there was
a thread change in between the bump-alloc which raised the level
and the sweeper examining the newly bumped space, the sweeper
would read uninitialized memory.

Explicitly initialize the memory before releasing the old gen lock, so
that this case won't happen. `FreelistCell` will be ignored by the sweeper.
If the mutator resumes and initializes it to a non-free-list-cell, it'll have its mark bit
set, which means the sweeper will also ignore it for that reason.

Reviewed By: davedets

Differential Revision: D22417639

fbshipit-source-id: 39397f901c6565ab6bd0b51e378fd9f20b820f81