Inline splitInternal for RegExp.propotype[@@split]
Summary:
`splitInternal` is getting very complicated, especially since it
handles splitting on both strings and regex. This also makes its
behaviour difficult to verify against the spec since the spec
provides separate implementations for the string and regex versions.
This diff copies and inlines `splitInternal` in the implementation
for `RegExp.propotype[@split]` with minimal modification. A
subsequent change will then bring the implementation more closely in
line with the spec.

Overall, this change should also improve performance since there is
no need to keep checking and casting the separator.

Reviewed By: avp

Differential Revision: D22395240

fbshipit-source-id: 16a92e9ff154b69410a1d48d5a23a64b54c34fbc