Change the Hades background thread to be joinable
Summary:
Instead of detaching the background thread and using a condition variable
to wait for its completion, the thread can be joined.
This cuts down on using additional data structures and thread messaging,
and uses simpler primitives. It also gives stronger guarantees that there will
never be multiple background threads.

Reviewed By: davedets

Differential Revision: D22574292

fbshipit-source-id: 940e6f80543c7ac55c59acefed05f84e0c94ee2c