Add support for Handle-SAN to Hades
Summary:
Handle-SAN is a build mode of Hermes that causes random collections to occur,
with the objective of moving the heap and causing raw pointer errors to show up.
It's also a good way to stress-test the system to be able to handle a lot of collections.

Enable this mode for Hades by checking the sampled flag (`shouldSanitizeHandles()`),
and then forcing an OG collection to start, if one is not already in progress.

This is done by adding a new `forceOldGenCollection` parameter to `youngGenCollection`.
Also, refactor `collect()` to use that same parameter to reduce some locking and
unnecessary checks.

Reviewed By: davedets

Differential Revision: D22400102

fbshipit-source-id: f76ed99d1dee566fc505af0f4fe27dd6ec7d7568