Use new constructor in RegExp.prototype[@@split]
Summary:
Now that there is a way to copy RegExp objects without performing a
full recompilation, use that. This is technically a little slower
since we need to construct a new object and copy the bytecode over.
However, it simplifies modifying the sticky flag since there is no
user visible modification any more.

Reviewed By: avp

Differential Revision: D22731730

fbshipit-source-id: c927b140d62c862e88a27bbad9698f7858d40f04