Change MarkWorklist to acquire locks less frequently
Summary:
Before this change, every call to `MarkWorklist::enqueue` from the
mutator write barrier would acquire a lock. This can be improved
by only locking in "chunks".

A chunk is a small fixed-size buffer that is filled up by the write barrier.
When the chunk is filled, the lock is acquired, and the chunk's contents
are moved. The marking thread reads off the pull chunks after acquiring its lock.
This means the lock is acquired by the mutator at most 1 in every N write barriers,
where N is the chunk size.

In order to ensure the push chunk is always considered before completing
marking, during the stop-the-world pause the maybe-not-full chunk is flushed out to
the marking thread. This is cheap since the world is already stopped.

This optimization yields significant improvements on some benchmarks that
have a lot of write barriers. An example is Splay.

Reviewed By: neildhar

Differential Revision: D22212576

fbshipit-source-id: ae24bc9e440d1163fed4c5f016832526c65de1ef