Add compressed pointers to Hades
Summary:
Compressed pointers is a build mode of Hermes where pointers only
take 32 bits on a 64-bit architecture device. This can shrink the heap size
by about 20%.

In order to support compressed pointers in Hades, it needs to store
a segment number as the first few bytes of any segment.
`PointerBase` uses this number to do the pointer conversion.

Reviewed By: kodafb

Differential Revision: D21913302

fbshipit-source-id: aadf238a76ccdf93cfd29fbb334593ef876e3c47