In a 32-bit build of Hades, interleave OG work with YG collections
Summary:
Hades can't do concurrent marking on a 32-bit architecture device,
since those devices don't allow a way to atomically mutate a
64-bit HermesValue. But Hades also has to run on 32-bit devices in
order to be useful for the large amount of older Android phones
out in the world.

So, Hades will now work in an incremental mode for a 32-bit build.
It will do this with mostly the same code, except all locks are replaced
with `FakeMutex`, and all atomics are replaced with `FakeAtomic`.

At each Young Gen collection, it'll check if there is still work to be
completed, and then drain the mark worklist in a small increment.
Once the mark stack is drained, it'll block for a bit and do sweeping
work to finish the collection. A future improvement could make
sweeping incremental as well.

Reviewed By: davedets

Differential Revision: D21791717

fbshipit-source-id: 099209a7761832e5182b9fa940807028da3961e3