Fix UAP in parseESDate
Summary:
In the case where input string contains a lot of trailing spaces,
iterator reaches the end of string in `consumeSpaces()`. But later the
iterator is used at `*it` in the TZ parsing.

This diff fix this potential UAP by moving `consumeSpace()` before
the end check.

This diff also correct a comment.

Reviewed By: avp

Differential Revision: D23084218

fbshipit-source-id: 27ad4ee9b269929917394f1206b66fe2a6b23579