Use GetMethod in yield*.
Summary:
`yield*` is supposed to use GetMethod to retrieve the `return` and `throw`
functions from the target iterator. In particular, GetMethod collapses
`null` and `undefined` both to `undefined`, which is relied on by
subsequent lines in the spec.

Add a `getMethod` builtin to allow this without increasing code
size too much.

Reviewed By: tmikov

Differential Revision: D22101041

fbshipit-source-id: bb34f25b14f8059cdf808c781e3270818173a939