Make Array.prototype.sort stable with index array
Summary:
closes #212

The current version of Array.prototype.sort uses QuickSort, which is not stable and thus not compliant with spec.

This diff uses an index array to keep items in stable order by comparing indices when items are equal. The tradeoff is the extra O(n) space for the index array.

The less() function of the SortModel has been updated to a compare() function to check for equality more easily.

For the future: Experiment with TimSort, which is better when the input is partially sorted. Note that the SortModel only allows swapping and doesn't allow allocating space for items, so everything has to be in-place, which makes implementing TimSort tricky because the merge operation has to be in-place. (A theoretically good but fairly complex in-place merge sort is Block sort/Wiki sort. There also exists a simpler but slightly worse alternative with O(n log^2n) worst case time). We could also just modify SortModel to allow for allocation.

Some rough benchmark on integer arrays for the new version, with comparisons to current version:

```
Size        | 10,000                     | 100,000                     | 1,000,000                   | 10,000,000
------------|----------------------------|-----------------------------|-----------------------------|---------------------
Random      | 22ms (5% slower)           | 282ms (6% slower)           | 3085ms (2% faster)          | 36218ms (<1% slower)
Sorted      | 13ms (same)                | 172ms (same)                | 2024ms (<1% slower)         | 23756ms (2% faster)
Same        | 14ms (22% faster)          | 172ms (29% faster)          | 2046ms (28% faster)         | 23610ms (32% faster)
```

Note:
1. 'Same' means all elements are the same. This, along with 'Sorted', serves as an edge case for quicksort.
2. The benchmark is run 3 times. Sorted and Same are deterministic, so the fastest times are taken. For Random, the median is used.
3. The time is wall time, not CPU time (as I said, rough benchmarks XD).
4. If a compare function is not provided to sort(), the input gets treated as string. In that case, the new version is actually slightly faster thanks to the optimization made to accomodate equality checking.

Reviewed By: dulinriley

Differential Revision: D23088088

fbshipit-source-id: 0170e4777f87132cdb579a158b6972733b5fe144