Move the non-concurrent GC work drain to yieldToOldGen
Summary:
Previously, there was a check during YG collection to see if some OG work was
pending, and to drain some of that work, so that OG eventually completes.
This type of periodic question "is OG done?" is the same question
`yieldToOldGen` asks, so it makes more sense to group them
together.

Also, rename `yieldToBackgroundThread` to `yieldToOldGen` because it
now does something even without a background thread active.

Reviewed By: neildhar

Differential Revision: D23245940

fbshipit-source-id: a9a5f81f703c0ec93b3e07c524989a56f6d6b8bc