Basic structure for parser compiled to WASM
Summary:
Set up initial structure for Hermes parser compiled to WASM. Results from parsing are returned in a ParseResults object which must be freed by the JS caller. This ParseResults object contains a fatal parse error if one exists (note that placeholder messages are used for now, returning the proper parse error will be implemented soon), as well a serialized AST. For simplicity this diff uses the `ESTreeJSONDumper` to serialize the AST to JSON so it can be deserialzed into proper JS objects. This is not very efficient and we will do better soon, but this gives us an AST in JS that we can work with for now.

Compiling with emscripten will produce a `hermes-parser-wasm.js` file that we can wrap in `HermesParser.js` to supply a clean API. `HermesParser.js` handles copying the source code to the WASM heap and calling the WASM parser module, then extracting the results, freeing memory, and either throwing an error or returning the AST.

This will be iterated upon, for now lets just set up the basic structure.

Reviewed By: avp

Differential Revision: D23323441

fbshipit-source-id: 0482bd3a82494644f344a7f7df87ac5225bfb572