Build JS AST by walking Hermes AST and using JS library builder functions
Summary:
The Hermes parser compiled to WASM can efficiently parse into its own internal memory representation of the AST. However this is not accessible from JS, so we must convert the entire AST into true JS objects that can be used by the rest of the program. The first iteration simply serialized the AST to JSON within the Hermes parser and then deserialized in JS to create a prop AST in JS. This is quite inefficient though, and we can do better.

This diff introduces a new `HermesParserJSBuilder` utility which walks the Hermes AST and builds up the AST in JS during this traversal using a new library of JS builder functions. We must link the new `HermesParserJSLibrary.js` for use in emscripten to provide the implementation of these builder functions. The `HermesParserJSLibrary` exposes a `build_<AST NODE>` function for each node in the Hermes AST (as defined in `ESTree.def`), along with a handful of other functions for building JS arrays and objects. The signatures of these external JS functions are then defined in `HermesParserJSLibrary.h` using the ESTree macros. This diff supplies a manual implementation of all the builder functions which I quickly put together based on Hermes' ESTree definitions, but in the future it would be great to switch these to be manually generated.

Complex values like JS objects and JS arrays cannot be passed as arguments to functions called between JS/WASM. Instead, we must generate JS objects and arrays from within JS and store them in a map indexed by a numeric ID. We refer to these as IDs as "JS references ids", and these ids can be passed as arguments and returned from functions in JS/WASM interop. The JS library has a simple `JSReferences` API for handling this. We supply a `store` option which takes in a JS value and adds it to the references map, returning a new JS reference ID. We also supply a `pop` option which takes in a JS reference ID, removes it from the map, and returns the referenced JS value. As long as we maintain the invariant that there is exactly one `pop` corresponding to every `store` then this map will not have "memory leaks" where values are unable to be garbage collected due to a lingering reference in this map (this is maintained in practice since every node is referenced exactly once, by its parent, so we must only manually release the top level program node).

Now for each AST node we use macros to generate an appropriate `build` function which returns a `JSReference` id for the newly created JS object node. These functions pass a `JSReference` to the node's loc, as well as all fields for that AST node. Each field type is handled as follows:

- Numbers can be passed directly as arguments, no special handling required.
- Booleans can be passed directly as arguments for functions. On the JS side we must wrap them in a call to `Boolean`.
- Strings can be sent to JS by passing in the `char *` to a null-terminated string. On the JS side we then must use `UTF8ToString` to extract the string itself. This is encapsulated in the `getString` helper method in the JS library
- Objects will be passed as their `JSReference` handle
- Arrays will be passed as their `JSReference` handle. We construct arrays using the `buildArray` and `appendToArray` JS library functions, though in the future we can likely optimize this to send a pointer+size to a contiguous array of `JSReferences` on the WASM heap.

Currently we are simply focused on getting all the data for the AST into JS. We will still need to tweak the format of the AST a bit to be strictly compliant with ESTree (as well as fix the structure of source locations, whose structure is currently temporary just to get the data into JS).

Overall, this has much better performance than the naive method of serializing/deserializing from JSON.

This also required moving the entire CMake build behind the `EMSCRIPTEN` flag so that it only compiles for WASM builds. Without this we get link errors for the functions defined in the JS library.

Reviewed By: avp

Differential Revision: D23333454

fbshipit-source-id: 1ec216508d498d49b224989935bedaf15bb0efdf