Avoid duplicating source map URL for every lazy function
Summary:
In lazy mode, we store a copy of the filename and source map URL in each
generated module. This is a problem when the source map URL is a large literal
data string.

Since we only need one copy, we can simply skip encoding it for any
lazily compiled module, and just get it from the parent module instead.

On one test file, this reduced peak RSS from 713MB to 90MB. This is compared
to 30MB without a large URL.

We still keep several copies around somewhat unnecessarily, but they're harder
to deduplicate. These include a copy in the SourceErrorManager, a copy with a
zero terminator for uniquing, one in the BytecodeModule string table, and
other related bookeeping.

Reviewed By: tmikov

Differential Revision: D23144986

fbshipit-source-id: 2cf95d8c155cc4b4df8877ecd2f5ba08159af12d