Remove HERMESVM_USE_JS_LIBRARY_IMPLEMENTATION
Summary:
`HERMESVM_USE_JS_LIBRARY_IMPLEMENTATION` is an experiment of using JavaScript
to implement part of the JS standard libraries (a.k.a JSLib in Hermes).

This diff removed this experiment since it wasn't able to provide a competitive
performance while adding some code complexities.

All the TODOs introduced in this diff serve as a merge helper and are resolved
in the later diff that makes internal bytecode always execute.

Reviewed By: dulinriley

Differential Revision: D23383872

fbshipit-source-id: 117941a2e2bf7d0d86d7991d7c068e067dda6116